// offscreen_service/src/types.ts
export const NATIVE_TOKEN_PROGRAM_ID = "credits.aleo";

export const ALPHA_TOKEN_PROGRAM_ID = "alphaswap.aleo";

export const NATIVE_TOKEN_TOKEN_ID = "__NATIVE_TOKEN__";
// ...
export interface AleoSendTxParams {
  privateKey: string;
  address: string;
  localId: string;
  chainId: string;
  programId: string;
  functionName: string;
  inputs: string[];
  baseFee: string;
  priorityFee: string;
  feeRecord: string | null;
  timestamp: number;
  amount?: string;
  tokenId: string;
}
// ...
export type Fee = {
  transition: Transition;
  global_state_root: string;
  proof: string;
};

export type Execution = {
  edition: number;
  transitions?: Transition[];
};

export type Transaction = {
  type: string;
  id: string;
  execution: Execution;
  fee?: Fee;
};

export type AleoTransaction = Transaction;
// ...
export type AleoLocalTxInfo = Omit<
  AleoSendTxParams,
  "privateKey" | "chainId"
> & {
  error?: string;
  status: AleoTxStatus;
  timestamp: number;
  transaction?: AleoTransaction;
  notification: boolean;
};

export type AleoTxWithTime = Transaction & {
  status: AleoTxStatus;
  timestamp: number;
};

export enum AleoTxStatus {
  QUEUED = "Queued",
  GENERATING_PROVER_FILES = "Generating Prover Files",
  GENERATING_TRANSACTION = "Generating Transaction",
  BROADCASTING = "Broadcasting",
  COMPLETED = "Completed",
  REJECTED = "Rejected",
  FINALIZD = "Finalized",
  FAILED = "Failed",
  UNACCEPTED = "Unaccepted",
}
// ...
export interface AleoRequestDeploymentParams {
  privateKey: string;
  chainId: string;
  address: string;
  localId: string;
  program: string;
  programId: string;
  baseFee: string;
  priorityFee: string;
  feeRecord: string | null;
  timestamp: number;
}